--- blog/admin.py
@@ -1,9 +1,0 @@
-from django.contrib import admin
-
-from .models import Post
-
-
-@admin.register(Post)
-class PostAdmin(admin.ModelAdmin):
-    list_display = ("title", "author", "created_at", "image")
-    prepopulated_fields = {"slug": ("title",)}
