+++ website/templates/company/bughunt/view_bughunt.html
