+++ website/templates/company/register_company.html
